requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
decorator==5.1.1
Django==4.0.3
future==0.18.2
geocoder==1.38.1
gunicorn==20.1.0
idna==3.3
iso8601==1.0.2
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
pyserial==3.5
PyYAML==6.0
pyzbar==0.1.9
qrcode==7.3.1
ratelim==0.1.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
what3words==3.1.1
whitenoise==6.0.0
windows-curses==2.3.0
